REAFACTOR: Se refactorizan estilos y nombre de los archivos
