--- src/features/usuarios/components/Detalles/detalles.module.css
@@ -1,17 +1,0 @@
-/* Contenedor */
-.box {
-  background: white;
-  padding: 20px;
-  border-radius: 8px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-  width: 320px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-/* texto para que se vea un poco mejor */
-.texts {
-  color: #333;
-  font-weight: 600;
-  font-size: 20px;
-}
