--- src/features/usuarios/components/Detalles/editUser.module.css
@@ -1,71 +1,0 @@
-.overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.modal {
-  background: #fff;
-  padding: 2rem;
-  border-radius: 12px;
-  box-shadow: 0 8px 20px rgba(0, 0, 0, 0.2);
-  width: 90%;
-  max-width: 500px;
-}
-
-.form label {
-  display: block;
-  margin-bottom: 0.8rem;
-  font-weight: 600;
-}
-
-.form input {
-  width: 100%;
-  padding: 0.5rem;
-  margin-top: 0.3rem;
-  margin-bottom: 1rem;
-  border: 1px solid #ccc;
-  border-radius: 6px;
-}
-
-.buttons {
-  display: flex;
-  justify-content: flex-end;
-  gap: 1rem;
-}
-
-.cancel {
-  background: #e74c3c;
-  color: white;
-  padding: 0.5rem 1rem;
-  border: none;
-  border-radius: 6px;
-  cursor: pointer;
-}
-
-.save {
-  background: #27ae60;
-  color: white;
-  padding: 0.5rem 1rem;
-  border: none;
-  border-radius: 6px;
-  cursor: pointer;
-}
-/* ---------------------------------------- */
-/* Estilos para el selector de rol y estado */
-/* ---------------------------------------- */
-.rolSelector {
-  margin-top: 1rem;
-  margin-bottom: 1rem;
-}
-
-.botonesRol {
-  display: flex;
-  gap: 0.5rem;
-}
